ci: update labgrid tests, add numbering
Python does not allow redefinitions of functions, or worse, it is
accepted and having twice a function by the same name ends up with
quite weird behavior and ununderstandable error messages. The
language does not prohibit to do so, thus the fix is adding a
rough incrementing numbering to the test_ function names.

Signed-off-by: Lothar Rubusch <l.rubusch@gmail.com>